Change some long to intptr_t to fix crash on 64-bit Windows. long is 64 bits on Linux but 32 bits on Windows. Some pointer values were being set via a long causing pointers to be truncated.

+++ src/Libraries/2D/Source/ctxmac.h
@@ -98,7 +98,7 @@ do {                            \
 #define gr_get_fill_type() (grd_canvas->gc.fill_type)
 
 #define gr_set_fill_parm(parm) \
-   (grd_canvas->gc.fill_parm=(long)(parm))
+   (grd_canvas->gc.fill_parm=(intptr_t)(parm))
 #define gr_get_fill_parm() (grd_canvas->gc.fill_parm)
 
 /* macros for setting the clipping region of a specified canvas. */
